+++ R/combo.R
@@ -45,112 +45,6 @@ convert_combo_data_to_dt <-
   }
 
 
-#' get_iso_colors
-#' 
-#' 
-#' @param  normalization_type charvec normalization_types expected in the data
-#' @keywords combination_data
-#'
-#' @return named charvec with iso colors
-#' 
-#' @examples 
-#' get_iso_colors()
-#' 
-#' @export
-get_iso_colors <-
-  function(normalization_type = c("RV", "GR")) {
-    normalization_type <- match.arg(normalization_type)
-    iso_cutoff <- seq(0, 1, 0.05)
-    breaks <- iso_cutoff
-    if (normalization_type == "GR") {
-      colors <- vapply(iso_cutoff, function(x) {
-        color_vector <- c(70, round((0.85 - x * 0.7) * 170), round((1.1 - x * 0.7) * 200))
-        assert_RGB_format(color_vector)
-        sprintf("#%s", paste(as.hexmode(color_vector), collapse = ""))
-      },
-      character(1)
-      )
-    } else {
-      colors <- vapply(iso_cutoff, function(x) {
-        color_vector <- c(70, round((1 - x * .85) * 170), round((1.1 - x * .85) * 232))
-        assert_RGB_format(color_vector)
-        sprintf("#%s", paste(as.hexmode(color_vector), collapse = ""))
-      },
-      character(1)
-      )
-    }
-    names(colors) <- iso_cutoff
-    colors
-  }
-
-assert_RGB_format <- function(x) {
-  if (any(x > 255)) {
-    stop("Some value is greater than 255. Not valid RGB format.")
-  }
-}
-
-#' Get colorscale data for given combo assay and growth metric
-#'
-#' @param g_metric growth metric
-#' @param assay_type assay type
-#' @keywords combination_data
-#'
-#' @return list with colors, breaks and limits
-#' @examples 
-#' get_combo_col_settings("GR", "excess")
-#' 
-#' @export
-get_combo_col_settings <-
-  function(g_metric,
-           assay_type) {
-    assay_names <- get_combo_assay_names()
-    assay_types <- names(assay_names)
-    checkmate::assert_choice(g_metric, c("RV", "GR"))
-    checkmate::assert_choice(assay_type, assay_types)
-
-    colors <- breaks <- limits <- NULL
-    if (assay_type %in% names(get_combo_assay_names(group = "combo_iso"))) {
-      myv <- get_iso_colors(g_metric)
-      colors <- as.character(myv)
-      breaks <- names(myv)
-    } else if (assay_type %in% c(names(get_combo_assay_names(group = "combo_excess")),
-                                 names(get_combo_assay_names(group = "combo_score")))) {
-      colors <- c("#003355", "#4488dd", "#eeeedd", "#CC8844", "#662200")
-      if (g_metric == "GR") {
-        breaks <- c(-0.5, -0.25, 0, 0.25, 0.5)
-      } else if (g_metric == "RV") {
-        breaks <- c(-0.3, -0.15, 0, 0.15, 0.3)
-      } else {
-        stop(sprintf("unexpected 'g_metric' type: '%s'", g_metric))
-      }
-    } else if (assay_type %in% names(get_combo_assay_names(group = "combo_score_mx"))) {
-      colors <- c("#003355", "#4488dd", "#eeeedd", "#CC8844", "#662200")
-      breaks <- c(-4, -2, 0, 2, 4)
-    } else if (assay_type %in% names(get_combo_assay_names(group = "combo_base_mx"))) {
-      if (g_metric == "GR") {
-        colors <- c("#001155", "#1122AA", "#AA4400", "#FF7711", "#ffffee")
-        breaks <- c(-0.6, -0.2, 0.2, 0.6, 1)
-      } else if (g_metric == "RV") {
-        colors <- c("#330033", "#770033", "#BB6633", "#ffaa11", "#ffffee")
-        breaks <- c(0, 0.25, 0.5, 0.75, 1)
-      } else {
-        stop(sprintf("unexpected 'g_metric' type: '%s'", g_metric))
-      }
-    } else {
-      stop(sprintf("no logic found for the assay_type: '%s'", assay_type))
-    }
-    stopifnot(
-      "unexpected error when determining combo color settings - either 'colors' or 'breaks' is NULL" = 
-        !(is.null(colors) || is.null(breaks))
-    )
-    list(
-      colors = colors,
-      breaks = breaks,
-      limits = c(min(breaks), max(breaks))
-    )
-  }
-
-
 DATA_COMBO_INFO_TBL <- data.table::data.table(
   name = c("hsa_score", "bliss_score", "CIScore_50", "CIScore_80",
            "smooth", "hsa_excess", "bliss_excess"),
@@ -261,7 +155,6 @@ define_matrix_grid_positions <- function(conc1, conc2) {
   
   list(axis_1 = axis_1, axis_2 = axis_2)
 }
-
 
 #' Round concentration to ndigit significant digits
 #'